# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_domain_site.CreateRequest:
    properties:
      description:
        type: string
      domain:
        type: string
      listen_port:
        type: integer
      name:
        type: string
      ssl_cert:
        type: string
      ssl_enabled:
        type: boolean
      ssl_key:
        type: string
      upstream:
        items:
          type: integer
        type: array
      waf_enabled:
        type: boolean
    type: object
  internal_domain_site.Site:
    properties:
      created_at:
        type: string
      description:
        type: string
      domain:
        type: string
      id:
        type: integer
      listen_port:
        type: integer
      name:
        type: string
      ssl_cert:
        type: string
      ssl_enabled:
        type: boolean
      ssl_key:
        type: string
      status:
        type: string
      updated_at:
        type: string
      upstream:
        items:
          type: integer
        type: array
      waf_enabled:
        type: boolean
    type: object
  internal_domain_site.UpdateRequest:
    properties:
      description:
        type: string
      domain:
        type: string
      listen_port:
        type: integer
      name:
        type: string
      ssl_cert:
        type: string
      ssl_enabled:
        type: boolean
      ssl_key:
        type: string
      status:
        type: string
      upstream:
        items:
          type: integer
        type: array
      waf_enabled:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API 支持
  description: WAF 防火墙管理系统控制面 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WAF Control API
  version: "1.0"
paths:
...
  /sites:
    get:
      consumes:
      - application/json
      description: 分页查询站点列表，支持按状态和关键字筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 状态筛选
        in: query
        name: status
        type: string
      - description: 搜索关键字
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 站点列表，包含 data、total、page、size 字段
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取站点列表
      tags:
      - 站点管理
    post:
      consumes:
      - application/json
      description: 创建新的防护站点，需提供站点名称和域名
      parameters:
      - description: 站点创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_site.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_domain_site.Site'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建站点
      tags:
      - 站点管理
  /sites/{id}:
    delete:
      consumes:
      - application/json
      description: 根据站点 ID 删除站点
      parameters:
      - description: 站点 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 站点不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除站点
      tags:
      - 站点管理
    get:
      consumes:
      - application/json
      description: 根据站点 ID 获取站点详细信息
      parameters:
      - description: 站点 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 站点详情
          schema:
            $ref: '#/definitions/internal_domain_site.Site'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 站点不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取站点详情
      tags:
      - 站点管理
    put:
      consumes:
      - application/json
      description: 根据站点 ID 更新站点信息，支持部分字段更新
      parameters:
      - description: 站点 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 站点更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_site.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/internal_domain_site.Site'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新站点
      tags:
      - 站点管理
  /sites/{id}/devices:
    get:
      consumes:
      - application/json
      description: 根据站点 ID 获取该站点绑定的所有设备 ID 列表
      parameters:
      - description: 站点 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设备 ID 列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取站点绑定的设备列表
      tags:
      - 站点管理
    post:
      consumes:
      - application/json
      description: 将指定设备绑定到站点，建立站点与设备的关联关系
      parameters:
      - description: 站点 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 绑定参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 绑定设备到站点
      tags:
      - 站点管理
  /sites/{id}/devices/{deviceId}:
    delete:
      consumes:
      - application/json
      description: 解除站点与指定设备的绑定关系
      parameters:
      - description: 站点 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 设备 ID
        in: path
        name: deviceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 解绑成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 解绑站点设备
      tags:
      - 站点管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"